+++ src/test/java/nom/bdezonia/zorbage/type/data/point/TestPointAlgebra.java
@@ -116,7 +116,7 @@ public class TestPointAlgebra {
 		assertEquals(1.4*p.component(2), p3.component(2), 0);
 		
 		G.POINT.zero().call(p3);
-		G.POINT.scaleByRational().call(new RationalMember(BigInteger.valueOf(14),BigInteger.valueOf(10)), p, p3);;
+		G.POINT.scaleByRational().call(new RationalMember(14,10), p, p3);;
 		assertEquals(3, p3.numDimensions());
 		assertEquals(1.4*p.component(0), p3.component(0), 0);
 		assertEquals(1.4*p.component(1), p3.component(1), 0);
